add wrapping key cache to operations tests
